+++ pages/index.tsx
@@ -5,17 +5,13 @@ import type { NextPage } from "next";
 import { signIn } from "next-auth/react";
 import Head from "next/head";
 import { useRouter } from "next/router";
-import React, { useEffect, useState, type ReactNode } from "react";
+import { useState } from "react";
 import {
   Alert,
-  Badge,
   Button,
   Card,
   Col,
-  Dropdown,
-  DropdownButton,
   Form,
-  Image,
   OverlayTrigger,
   Placeholder,
   Row,
@@ -24,6 +20,12 @@ import {
 import useSWR from "swr";
 import type { UserV2 } from "twitter-api-v2";
 import Layout from "../components/Layout";
+import ProfileSummary, {
+  TwitterProfileName,
+  TwitterProfileSummary,
+} from "../components/ProfileSummary";
+import TasklistPicker from "../components/TasklistPicker";
+import useTasks from "../hooks/useTasks";
 
 const NameSampleComponent: React.FC<{
   color: string;
@@ -60,54 +62,6 @@ const NameSampleComponent: React.FC<{
 
 const dummyTasks = ["タスクその1", "タスクその2", "タスクその3"].map(
   (title, id) => ({ id, title })
-);
-const ProfileSummary: React.FC<{
-  id: string;
-  name: ReactNode;
-  image: string;
-}> = ({ id, name, image }) => (
-  <>
-    <Row className="gx-3">
-      <Col xs="auto">
-        <Image
-          width={48}
-          height={48}
-          src={image}
-          alt="avatar photo"
-          roundedCircle
-        />
-      </Col>
-      <Col>
-        <div>{name}</div>
-        <div>
-          <small className="text-muted">{id}</small>
-        </div>
-      </Col>
-    </Row>
-  </>
-);
-
-const TwitterProfileName: React.FC<{
-  name: ReactNode;
-  isProtected: boolean;
-}> = ({ name, isProtected }) =>
-  isProtected ? (
-    <>
-      {name} <i className="bi bi-lock-fill" />
-    </>
-  ) : (
-    <>{name}</>
-  );
-
-const TwitterProfileSummary: React.FC<{
-  user: Omit<TwitterUser, "name">;
-  name: ReactNode;
-}> = ({ user, name }) => (
-  <ProfileSummary
-    name={<TwitterProfileName name={name} isProtected={user.protected} />}
-    id={`@${user.username}`}
-    image={user.profile_image_url}
-  />
 );
 
 const beginningTextTitle = "先頭テキスト";
@@ -282,125 +236,6 @@ const NameComponentInput: React.FC<{
         />
       </Col>
     </Form.Group>
-  );
-};
-
-const useTasks = (tasklist: string | null | undefined) => {
-  const { data } = useSWR<readonly tasks_v1.Schema$Task[]>(
-    tasklist && `/api/tasklists/${tasklist}/tasks`,
-    fetcher
-  );
-  return data;
-};
-
-const TasklistPicker: React.FC<{
-  tasklist: string | undefined;
-  onChange: (tasklist: string) => void;
-}> = ({ tasklist: tasklistId, onChange }) => {
-  const { data: tasklists } = useSWR<readonly tasks_v1.Schema$TaskList[]>(
-    "/api/tasklists",
-    fetcher
-  );
-
-  const defaultTasklist = tasklists?.[0];
-  const tasklist = tasklists?.find((tasklist) => tasklist.id === tasklistId);
-
-  useEffect(() => {
-    if (!tasklist && defaultTasklist) {
-      onChange(defaultTasklist.id!);
-    }
-  }, [defaultTasklist, onChange, tasklist]);
-
-  const tasks = useTasks(tasklist?.id);
-
-  return (
-    <Card body>
-      <Row>
-        {tasklist ? (
-          <Col>
-            <Card.Title>
-              <i className="bi bi-list-task" /> {tasklist.title}
-            </Card.Title>
-            <Card.Text>
-              <small className="text-muted">
-                {new Date(tasklist.updated!).toLocaleString(undefined, {
-                  dateStyle: "short",
-                  timeStyle: "short",
-                })}
-              </small>
-            </Card.Text>
-          </Col>
-        ) : (
-          <Col xs={9} xl={10}>
-            <Placeholder as={Card.Title} animation="glow">
-              <Placeholder xs={6} />
-            </Placeholder>
-            <Placeholder as={Card.Text} animation="glow">
-              <Placeholder xs={4} size="sm" />
-            </Placeholder>
-          </Col>
-        )}
-        {tasklists ? (
-          <Col xs="auto">
-            <DropdownButton
-              id="tasklist-dropdown-button"
-              title="変更"
-              variant="secondary"
-            >
-              {tasklists.map((tasklist) => (
-                <Dropdown.Item
-                  key={tasklist.id}
-                  onClick={() => onChange(tasklist.id!)}
-                  active={tasklist.id === tasklistId}
-                >
-                  <div>{tasklist.title}</div>
-                  <div>
-                    <small>
-                      {new Date(tasklist.updated!).toLocaleString(undefined, {
-                        dateStyle: "short",
-                        timeStyle: "short",
-                      })}
-                    </small>
-                  </div>
-                </Dropdown.Item>
-              ))}
-            </DropdownButton>
-          </Col>
-        ) : (
-          <Col xs={3} xl={2}>
-            <Placeholder.Button
-              variant="secondary"
-              xs={12}
-            ></Placeholder.Button>
-          </Col>
-        )}
-      </Row>
-      {tasks ? (
-        tasks.length > 0 ? (
-          <div>
-            {tasks
-              .map((task) => (
-                <Badge pill bg="dark" key={task.id}>
-                  {task.title}
-                </Badge>
-              ))
-              .reduce((previousValue, currentValue) => (
-                <>
-                  {previousValue} {currentValue}
-                </>
-              ))}
-          </div>
-        ) : (
-          <Card.Text>
-            <i>未完了のタスクはありません</i>
-          </Card.Text>
-        )
-      ) : (
-        <Placeholder as={Card.Text} animation="glow">
-          <Placeholder xs={2} /> <Placeholder xs={3} /> <Placeholder xs={2} />
-        </Placeholder>
-      )}
-    </Card>
   );
 };
